chore(ci): drop accumulated fmt and clippy drift, fix unused props in WorkspaceSwitcher
cargo fmt --all + cargo clippy --fix across the workspace, plus
targeted manual fixes for lints that --fix could not rewrite.

Crate-wide allows in lib.rs and main.rs for eight stylistic lints
(too_many_arguments, large_enum_variant, type_complexity,
should_implement_trait, doc_lazy_continuation,
doc_overindented_list_items, field_reassign_with_default,
manual_strip). Rationale comments in lib.rs cover each one.

Test-binary file-level allow on tests/dsn_corpus.rs because
integration tests are separate compilation units and the crate-
wide allows do not reach them.

Targeted code fixes:
- handlers/asset_audits.rs and handlers/asset_usage.rs: replace
  recorded_by.unwrap_or(Uuid::nil()) with auth.user_uuid; the
  preceding let binding makes recorded_by statically Some.
- services/plugins/manifest_validate.rs: convert an orphan ///
  doc block (the underlying constant was removed when the typed
  Permission enum landed) to // regular comments.
- services/plugins/provisioning.rs: regroup PROVISION_LOCK_KEY
  hex digits to 4-byte groups.
- services/email_queue/retry.rs: replace .min().max() chain with
  .clamp(1, MAX_BACKOFF_SECS).
- services/presence.rs: sort_by with Reverse closure becomes
  sort_by_key(|p| Reverse(p.last_active_at)).
- services/channels/pipeline.rs: two let-else returning None
  collapse to ? operators.
- handlers/msgraph_integration.rs: consolidate SMTP:/smtp: prefix
  strip with strip_prefix().or_else(strip_prefix(...)).
- handlers/mod.rs: collapse three identical /uploads/ branches
  into one trim_start_matches call.
- handlers/debug.rs: drop redundant "debug" | _ wildcard arm.
- handlers/documentation.rs, services/search/extractors.rs,
  utils/markdown_export.rs: match panic::catch_unwind { Ok(s) =>
  s, Err(_) => String::new() } becomes .unwrap_or_default().
- frontend/components/WorkspaceSwitcher.vue: drop the unused
  const props = withDefaults(...) capture; collapsed reads from
  the template-implicit prop.

ESLint, vue-tsc, cargo fmt --check, cargo clippy -D warnings,
cargo deny check, and npm audit all pass on the host before
this commit.

+++ backend/src/handlers/passkeys.rs
@@ -932,7 +932,7 @@ pub async fn start_passkey_setup_login(
                     "Account temporarily locked. Try again in {} seconds.",
                     remaining_seconds
                 ),
-                remaining_seconds as u64,
+                remaining_seconds,
             );
         }
         Ok(None) => {} // Not locked, continue
@@ -1110,7 +1110,7 @@ pub async fn finish_passkey_setup_login(
                     "Account temporarily locked. Try again in {} seconds.",
                     remaining_seconds
                 ),
-                remaining_seconds as u64,
+                remaining_seconds,
             );
         }
         Ok(None) => {} // Not locked, continue
